pyproject.toml
[project]
name = "stable-diffusion-webui-codex"
version = "0.0.0"
description = "Codex Stable Diffusion WebUI"
readme = "README.md"
requires-python = "==3.12.10"
dependencies = [
  # API server
  "fastapi==0.121.0",
  "starlette==0.49.1",
  "uvicorn==0.37.0",
  "pydantic==2.11.10",
  "httpx==0.28.1",
  "python-multipart==0.0.20",
  # Core utilities
  "requests==2.32.5",
  "rich==14.2.0",
  "tqdm==4.67.1",
  "psutil==7.1.0",
  "PyYAML==6.0.2",
  "piexif==1.1.3",
  # Model stack (Hugging Face / Diffusers)
  "accelerate==1.10.1",
  "diffusers==0.36.0",
  "transformers==4.57.3",
  "peft==0.17.1",
  "huggingface-hub==0.36.0",
  "tokenizers==0.22.1",
  "safetensors==0.7.0",
  "einops==0.8.1",
  "sentencepiece==0.2.1",
  # Imaging / numerics
  "numpy==1.26.4",
  "Pillow==10.4.0",
  # Upscalers / SR (Spandrel: unified loader/runner for ESRGAN/SwinIR/HAT/DAT/etc)
  "spandrel==0.4.1",
  # Small pins to keep resolver stable across platforms
  "filelock==3.20.3",
  "packaging==25.0",
  "regex==2025.9.18",
  "typing-extensions==4.15.0",
  # Windows-only niceties (optional in code, but improves console UX)
  "colorama==0.4.6; platform_system == 'Windows'",
  "protobuf==6.33.2",
]

[project.optional-dependencies]
# PyTorch is intentionally optional and selected by install scripts.
# Pick exactly one: cpu / cu126 / cu128 / cu130 / rocm64.
dev = [
  "pytest==8.3.4",
]
cpu = [
  "torch==2.9.1; sys_platform == 'darwin'",
  "torchvision==0.24.1; sys_platform == 'darwin'",
  "torch==2.9.1+cpu; sys_platform != 'darwin'",
  "torchvision==0.24.1; sys_platform == 'linux' and platform_machine == 'aarch64'",
  "torchvision==0.24.1+cpu; sys_platform != 'darwin' and platform_machine != 'aarch64'",
]
cu126 = [
  "torch==2.9.1; sys_platform == 'darwin'",
  "torchvision==0.24.1; sys_platform == 'darwin'",
  "torch==2.9.1+cu126; sys_platform != 'darwin'",
  "torchvision==0.24.1; sys_platform == 'linux' and platform_machine == 'aarch64'",
  "torchvision==0.24.1+cu126; sys_platform != 'darwin' and platform_machine != 'aarch64'",
]
cu128 = [
  "torch==2.9.1; sys_platform == 'darwin'",
  "torchvision==0.24.1; sys_platform == 'darwin'",
  "torch==2.9.1+cu128; sys_platform != 'darwin'",
  "torchvision==0.24.1; sys_platform == 'linux' and platform_machine == 'aarch64'",
  "torchvision==0.24.1+cu128; sys_platform != 'darwin' and platform_machine != 'aarch64'",
]
cu130 = [
  "torch==2.9.1; sys_platform == 'darwin'",
  "torchvision==0.24.1; sys_platform == 'darwin'",
  "torch==2.9.1+cu130; sys_platform != 'darwin'",
  "torchvision==0.24.1; sys_platform == 'linux' and platform_machine == 'aarch64'",
  "torchvision==0.24.1+cu130; sys_platform != 'darwin' and platform_machine != 'aarch64'",
]
rocm64 = [
  "torch==2.9.1+rocm6.4; sys_platform == 'linux'",
  "torchvision==0.24.1+rocm6.4; sys_platform == 'linux'",
  "pytorch-triton-rocm==3.5.1; sys_platform == 'linux'",
]

[tool.uv]
python-preference = "only-managed"
python-downloads = "manual"
conflicts = [
  [
    { extra = "cpu" },
    { extra = "cu126" },
    { extra = "cu128" },
    { extra = "cu130" },
    { extra = "rocm64" },
  ],
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[[tool.uv.index]]
name = "pytorch-rocm64"
url = "https://download.pytorch.org/whl/rocm6.4"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu", marker = "sys_platform != 'darwin'" },
  { index = "pytorch-cu126", extra = "cu126", marker = "sys_platform != 'darwin'" },
  { index = "pytorch-cu128", extra = "cu128", marker = "sys_platform != 'darwin'" },
  { index = "pytorch-cu130", extra = "cu130", marker = "sys_platform != 'darwin'" },
  { index = "pytorch-rocm64", extra = "rocm64", marker = "sys_platform == 'linux'" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "cpu", marker = "sys_platform != 'darwin'" },
  { index = "pytorch-cu126", extra = "cu126", marker = "sys_platform != 'darwin'" },
  { index = "pytorch-cu128", extra = "cu128", marker = "sys_platform != 'darwin'" },
  { index = "pytorch-cu130", extra = "cu130", marker = "sys_platform != 'darwin'" },
  { index = "pytorch-rocm64", extra = "rocm64", marker = "sys_platform == 'linux'" },
]
pytorch-triton-rocm = [
  { index = "pytorch-rocm64", extra = "rocm64", marker = "sys_platform == 'linux'" },
]
